soft-geo: separate soft-geo task and handler
